Scale model UI
Add a new "Not connected" warning for when it's not connected - this will improve the UX Significantly

+++ VTubeStudioApi/Requests/MoveModelRequest.cs
@@ -11,7 +11,7 @@ namespace Cazzar.StreamDeck.VTubeStudio.VTubeStudioApi.Requests
         [JsonProperty("positionX")] public double? PositionX { get; set; }
         [JsonProperty("positionY")] public double? PositionY { get; set; }
         [JsonProperty("rotation")] public double? Rotation { get; set; }
-        [JsonProperty("size")] public int? Size { get; set; }
+        [JsonProperty("size")] public double? Size { get; set; }
         
         public override RequestType MessageType { get; } = RequestType.MoveModelRequest;
     }
